Refactor AppNavigationContext to new style (#4)
* Refactor AppNavigationContext to new style

This change moves the NavigationProvider and Consumer into
AppNavigationContext so that they are more convenient to use. It also
refactors withNavigation to take a simple render prop to get at the
"navigation" object, and moves the old functionality into
AppNavigationContext.connect to be more canonical.

* Fix tabs

This change fixes some whitespacing issues.