Implement switch router (#3)
Squashing down the switch router into master. At this point, the code reviews are still too big for people to look at piecemeal, so we're just going to do a final review before we start using the new navigation system.

* Add stubs for stack and switch routers

* Add more validation checks for screen options

* Implement helpers and fix luacheck warnings

This change implements isValidRoactElementType and
getScreenForRouteName, while also fixing some existing luacheck errors.

* Implement getConfigGetter

This change implements the getConfigGetter builder function and its tests.
This builder creates a function that maps the hierarchy of navigation
options from multiple sources (route > component > navigator)) into a
flat list that can be used to properly configure the screen in question.

* Reorganize file structure for utils

This change moves utility files under a subfolder instead of keeping
them at top-level, since there are going to be a number of small
utilities we need that are outside the scope of routers/navigators/etc.

* Implement getChildRouter

This change implements getChildRouter, which searches a given router for
a child that can handle the provided routeName.

* Implement unique key generator for screens

This change implements a reasonable unique keygen so that we can tell
individual screens apat without risk of collision.

* Move CompleteTransition to NavigationActions

This change moves the complete transition stack action to
NavigationActions since it actually needs to be used by
SwitchRouter/TabRouter to support animations for those cases.
React-Navigation has a todo to relocate it in their codebase that they
haven't been able to fix yet, so let's just take care of it now...

* Implement validateRouteConfigMap

This change adds a function that checks a given route config
dictionary to make sure that all of the routes specified in it are
valid. (It does not check the output of getScreen() functions since
there would be a large performane cost for that).

* Add BackBehavior enum

This change adds a BackBehavior enum to be used with route configs that
need to customize what the back operation does.

* Implement getNavigationActionCreators

This function creates the goBack(), navigate(), and setParams()
actions that are made available through the navigation object
passed to UI components.

* Add getChildrenNavigationCache

This adds a getter that retrieves the child navigation cache from
the current navigation context, and also prunes unsupported keys
while it does so. The child navigation cache helps us avoid recreation
of components that we've created already.

* Add missing changes for NavigationEvents

* Add getNavigation

This change adds the getNavigation function, but its tests are not yet
complete because its dependency, getChildNavigation, is not finished
yet.

* First-cut implementation of SwitchRouter

This is the first cut of an implementation of SwitchRouter by
translating React-Navigation's SwitchRouter from JavaScript. From here,
we will see how we can clean up the code so it is more understandable.

* Add some of the unit tests for SwitchRouter

* Implement a bunch of router utils

This change pulls in a bunch of router utils that I implemented
originally in the implement_routers branch. I want to commit these for
code review separately from the routers so that it is easier for
reviewers to figure out what's going on. Therefor, these are
patch-applied into a new branch together for a single pull request to
mater.

* Rearrange some code into helpers

This change moves some of the easy-to-pull-out code into helper
functions so that the main function in SwitchRouter is easier to read.
More changes will follow later, after we have better unit test coverage.

* Add test to make sure child routers propagate

* Add switch router tests for navigation options

This change adds tests that check navigation options priority in switch
router to make sure that overrides are done in the correct order.

* Fix bug in createConfigGetter

This change fixes a bug in createConfigGetter around the configurer type
checks. It also adds unit tests to cover priority between default vs.
component vs. route-specified navigation options.

* Fix SwitchRouter child routing bug

There was a bug in SwitchRouter.collectChildRouters that would cause it
to error when a child screen was defined as a function instead of a
table. I added a type-check to avoid that.

Also added unit tests for getActionCreators and getComponentForState.

* Add test for SwitchRouter.getComponentForRouteName

* Implement init and back switchrouter tests

This change implements unit tests covering SwitchRouter's Init and Back
action handling. Still many more tests to add.

* Add unit tests for switchrouter navigate action

This change adds unit tests that validate SwitchRouter's response to
navigation actions, and the action param passed to child routers is
respected.

* Fix comments from pull request

Little bit of cleanup according to comments from pull request.
- Logic simplification
- KeyGenerator doesn't need to initialize random numbers system in
Roblox.

* Add test for switchrouter action setparams

This adds a test to switchrouter for setparams, a test that validates
switchrouter establishes its initial state if navigating to the
currently active child without any previous state, and fixes an error in
setparams handling.

* Add params init tests and fix nil table bug

This adds unit tests for params initialization and merging of
initialRouteParams with route-specific params. It also fixes a Cryo
merge bug that one of the tests found.

* Add switch router navigation state reset tests

This change adds tests to validate that SwitchRouter respects the
resetOnBlur setting to decide if it should clear a route's state to
initial state when navigating to other routes.

* More switchrouter tests

Adds tests for:
- params attached to init action
- making sure that active child gets first pass at handling a nav action

* Add final tests and clean up shallow copying

This change adds final tests to SwitchRouter to cover child params
merging and child getStateForAction failure. It also refactors the use
of Cryo.List.join() to do simple cloning and single index replacement
of lists because Cryo does not guarantee a unique list to be produced.
There is now a helper function for this purpose.

* Add shallowCopyDictionary to SwitchRouter

Just like we cannot rely on Cryo.List.join to return a new table every
time it is called, we have to implement a shallow copy for dictionaries
in the rare cases that we need to clone state. This adds that change,
and also makes the inputState checks for getStateForAction handle nil
more explicitly. We depend upon being able to nil-check the latter, so
if the user were to pass in some truthy value then things would explode.

* Remove todo comment for switchrouter tests

Tests are now complete enough to move on. Removed todo note.

* Add a stub skip test for stack router