Implement more tools and tests (#6)
* Add unit tests for createNavigator

This adds some basic unit tests for createNavigator().

* Add switch navigator test and needed fixes

This change adds a Roact mounting test for createSwitchNavigator and
fixes issues in getChildNavigation that blocked that from working.

I also changed the remove() calls for event listeners to be named
disconenct() to better match our traditional signaling API at Roblox.

* Implement TabRouter tests

This change adds some basic tests for TabRouter. Most functional cases
are covered by SwitchRouter, but we check that the adjusted defaults for
TabRouter work as expected.

* Add tests and fix issues in getChildEventSubscriber

This change adds many unit tests for getChildEventSubscriber and fixes
some code issues encountered while writing the tests. I also changed
the if/else logic for specific focus event handling to be a bunch of
executed "if" statements because that's what the actual React-Navigation
implementation does and I realized the disconnectAll() call has a
dependency on this behavior.

* Fix typo bug in getChildEventSubscriber

This change fixes a symbol naming typo that would have resulted in a
bug, which I found while writing the remainder of the tests for
getChildEventSubscriber.

* Add lots of tests for getNavigation

This adds the majority of the tests needed to cover getNavigation.

* Fix getChildNav.createParamsGetter and add tests

This change adds most of the unit tests for getChildNavigation and fixes
a performance+behavioral issue in the translation of createParamsGetter
from JS.

* Fix bugs in createAppContainer

This change fixes a typo in createAppContainer and adds the missing
implementation for didMount that triggers the initial navigation actions
that set everything up... Also adds tests to make sure that the
container registers with the back action signal.