This change fixes action event propagation (#15)
There was an issue in the propagation of events generated from
dispatched actions in AppContainer, where they would not be propagated
because the function that takes care of that was iterating over list
portion of the subscribers table, but the subscriber handlers were
stored in the dictionary portion.

Also cleaned up some niggling issues here and there.