Nested nav stories (#21)
* Fix child router bugs

This change fixes some bugs found in handling of child routers. It also
adds a story covering the nested stack router use case.

Bugs fixed:
1. Navigate action was being incorrectly processed by parent because it
fell through if child returned nil state (when it shouldn't process the
action).
2. BackAction handling did not account for NoneSymbol when passed as the
"from" key in the action.

* Add story for switch+stack navigator

This adds a story covering a switch navigator with a stack navigator as
a child.