Change other modules deps (#23)
This changes the module dependency for Cryo to point at the current
Roblox version to make sure our tests find compatibility problems. We
CANNOT revert to Otter/roblox because that version does not auto-start
motors and has timing bugs when you call start() repeatedly on the same
motor.

Roact-Navigation depends on Otter of at least this version:
9ad129e70e103d0de71232a0d0e7a1527da7a51a