Add more stories (#24)
* Add complex stack-in-switch-in-stack story

This change adds a story that demonstrates three levels of nested
routing.

* Add story for navigator embedded in component

This adds a story showing how to embed a navigator inside a component so
that it takes up a smaller visual area. I also added some more
complexity to the SwitchNavigatorWithSwitchChild story that shows how to
set up stack navigator titles for child navigators.

* Add sister navigators story

This change adds a story covering how to put two navigators on the same
screen as siblings. I also made the createAppContainer call more
stringent about what it allows since we depend upon being able to index
the router field.