Implement with navigation focus (#27)
* Initial cut of withNavigationFocus

This change implements first cut of withNavigationFocus and does some
cleanup.

* Fix isFocused logic and add story

This change implements a new story that checks out withNavigationFocus
and fixes a logic bug surrounding the navigation.isFocused() function.
It would return true so long as the current state key matched the screen
in question, but this is actually not quite right; we need to return
false while we're still _transitioning_ to that screen, even if we are
the active key. This results in the expected behavior of isFocused()
returning false when we haven't yet received the DidFocus event.