Public api cleanup (#37)
This change removes all the tab bar/bottom bar elements from the public RN API since we're going to be moving all of that into a separate repo that can take advantage of UIBlox. I also fixed a missing pass-through for StackActions.popToTop(), as well as adding missing tests and top bar public API elements.

(There was some carry-along work in a branch I was originally using to implement the bottom bar elements. This change is effectively merging back the useful things from that branch.)