Fix setstate timing in Transitioner layout (#38)
There was a bug in Transitioner's handling of layout changes. Since it
spawn()s a setState for the layout change, the application of that
change can land after an initial button click in a Rhodium test (or if
the app is doing something crazy at init time).

The setState implementation persisted the entire _old_ state, which had
a nasty side effect of obliterating any transitions that happened since
the spawn()ed routine was enqueued. This surfaced as an unexplained
unmount of PageTwo in the PushNav Rhodium test.

This change removes the setState since it is ostensibly safe in the
newest version of Roact to call setState synchronously from layout
change callbacks now. Cchange callbacks are accumulated and only run
AFTER all component mounting is finished now.)