Remove public top bar (#39)
* Remove top bar code

This change removes the existing TopBar code since we will no longer
include app-specific UI inside the roact-navigation core library. This
code will be re-created in another library that can take advantage of
UIBlox.

I also deleted now-unneeded helpers (EdgeInsets and
ContentHeightFitFrame).

Note that this change breaks most of the Storybooks. We will have to
update them to be more suitable to the new world.

* Update storybooks for no top bar

This change updates the storybooks to accommodate the fact that we've
removed the top bar from roact-navigation.

* Update Reset stack story to use barless nav

This change updates the new story added to RN to use the top-bar-less
createStackNavigator function instead of the missing one.

* Update storybooks to remove title option

This change updates the storybooks to remove navigationOptions.title,
which was used to configure the top bar title TextLabel. Since we no
longer have a top bar, it's not appropriate to include these properties
in storybooks for the core library.