Update RN to use new Roact 1.2.0
This change updates Roact Navigation to use Roact 1.2.0. Paul found an
issue which was introduced between RN 0.1.0 and 0.1.1 where we no
longer function correctly with switch navigators embedded inside stack
navigators. This is because we removed the spawn() calls for the async
portions of code after setState. We have to put these back until we get
an async setState callback handler in an even newer Roact.