Implement external transition callbacks (#45)
* Add transition callbacks and story

This change adds transition callbacks to StackNavigator and also adds a
new story to show how to exercise them. These transition callbacks can
be used to synchronize animation of any UI that is external to the stack
navigator.

* Modify stack nav test to check transition cbs

This change modifies the NavigationFlow test so that it also verifies
the correct call sequence of onTransitionStart/End/Step.