Migrate to registry deps (#47)
* Move available deps to registry

This change adjusts Roact, Cryo, and Otter to use registry dependencies.
StateTable is not available yet, pending some setup work by @pdoyle.
Once it is available, we will migrate that as well.

* Add changelog line for reg deps

This change adds a line to the changelog for 0.2.0 for registry
dependencies.

* Switch state-table to reg dep

This switches state-table to come from a registry dependency now that
it's available.

* Change rojo project structure and update readme

This change makes default.project.json compatible with building the rbxm
needed for rotriever registry dependencies. It also moves the original
default.project.json to storybook.project.json, updates RN naming
everywhere, and fixes the readme with instructiojns for how to build the
various moving parts in the new world.

* Update readme

Update the readme with instructions about rojo and horsecat setup
requirements, per @nukropina's review.