Alter tests to use direct top-level path (#48)
* Alter tests to use direct top-level path

There are two tests in RN that use top-level paths to get at public API
elements. We have to lop off ".Parent.RoactNavigation" on the end of
these imports so that the tests run correctly when being imported as a
dependency in a top-level folder that is not named "RoactNavigation".

* Update source files to use relative paths for root

The two source files need to use relative root so that they can
correctly reference when installed in a folder that is a different name
from the RoactNavigation import name we've been using.

I also changed bin/spec.lua to use roact-navigation as the package name
to better simulate the folder structure that is produced in our perforce
repo. This will make it much easier to detect these kinds of faults in
the future.