Support switchview keepmounted (#53)
* Allow SwitchView to keep all visited pages mounted

This provides a navigationConfig option for SwitchView so that it can
keep all previously visited pages mounted. Still need to update the
Rhodium tests!

* Update SwitchView Rhodium test

This change updates the Rhodium test for SwitchView so that it works in
the new schema.

* Add story for switch keep page mounted feature

This change adds a story that shows how to use SwitchNavigator/View's
keepVisitedPagesMounted feature.

* Update changelog

This updates the changelog to reflect the keepVisitedScreensMounted
feature.