Add generic external dispatch (#54)
* Replace backActionSignal with extDispatchConnector

THIS IS AN API BREAKING CHANGE!

This change removes the non-generic backActionSignal and replaces it
with a notion called "externalDispatchConnector" which allows
arbitrary NavigationActions to be injected into the top-level app
container element by means of a connector function.

RN subscribes to externally generated dispatch events by calling
externalDispatchConnector, and disconnects from it by calling its
result.

* Update changelog