Fix MOBLUAPP-2327: "Fix race condition in RN Transitioner" (#55)
* Fix MOBLUAPP-2327: "Fix race condition in RN Transitioner"

The Transitioner has an issue where it can sometimes receive a position
change callback after endTransition is executed, in complicated
environments like our LuaApp.

It turns out that this is due to a bugfix we carried over from React
Navigation where they were making up for an issue with the React
animation library (animations could be interrupted and the motor would
be left in the incorrect position).

We do not have this same problem in Roact Navigation with Otter, but
adopting their bugfix was causing an _extra_ completion callback to
fire. This completion callback would nil the _prevTransitionProps needed
in the _onPositionStep handler (else it throws an error).

We could have patched this in the step callback, but that would have
just masked the problem described above.

* Add protection for end transition call

This change adds a check for isTransitionRunning to ensure that the
endTransition call is always safe. I would have had to add an equivalent
comment explaining how didUpdate checks this already; it's just easier
and safer to write the code anyway.