Fix MOBLUAPP-2477: "Roact Navigation race condition when spamming nav actions"
There were a few issues:
1. The spawn() calls were inconsistent in Transitioner, leading to cases
where we would manually call endTransition synchronously, and later
spawned operations could run against the wrong state.
2. Otter motor onComplete could run endTransition against the wrong data
because another spawned call ran in the meantime.
3. Spawning inside Transitioner could interact with external app's
spawned operations in unpredictable ways (yet more race conditions). The
external code would be able to commit additional navigation actions
before the entire transition queue has been drained.
4. We were calling setState, and then spawning the follow-up ivar
changes to run later. This made the ivar changes potentially run against
incorrect Roact state.

It turns out, that we can:
a) Rely completely on the Otter position motor to trigger our transition
completion handling, negating the need for manually calling
endTransition except for the one case that does not use the motor.
b) (Because of A) get rid of all the spawn calls in Transitioner, making
the changes in calls to setState and all of the non-state based ivars
consistent.

Now the Transitioner code is very predictable and we're not seeing any
of the flickering/transition bustage that we had before this change.

I also took the opportunity to put a safety around the
_prevTransitionProps usage in the position motor callback. There are
cases where that ivar can be nil and we were hitting errors. The spawn
calls were originally put in the exact configuration they were to mask
these errors, but it really is valid for the Otter completion to happen
when prevTransitionProps is nil. (We set it to nil on endTransition when
no more transitions are available!)