Update Transitioner Otter spawns
This change removes the spawn call around the position step callback
since it's not necessary, reduces performance, and might lead to
timing problems for users of Roact Navigation.

I tried to remove the spawn around the completion handler, but this
leads to navigation lockups if you submit a nav action while a
transition is occurring, due to an Otter bug that I've noted in a
comment in Transitioner.lua. For now, we have to leave that spawn in
place.