Fix bug with absorbInput and add story
This change fixes a bug with navigationOptions.absorbInput where setting
it to false doesn't allow clicks to pass through on desktop platforms.
It works fine on mobile because touch input handling is treated
differently, however, there is a legacy issue that means we can't rely
on the same behavior on desktop. The fix is to switch over to a "Frame"
for the backing view for each stack card when absorbInput=false instead
of just using a TextButton with Active=false.