Fix flaky NavigationFlow rhodium test (#11)
The test was flaky because it does direct comparison of the list of
navigation event callbacks that were issued by RN on stack navigator
transition against a pre-defined sequence. There is no guarantee of
the order of any events fired within the same call stack because the
order of iterating on Lua tables is not fixed.

This commit changes the test so that it allows variation in the order of
the "WillFocus/WillBlur" and "DidFocus/DidBlur" event pairs internally,
but still requires that the "Will*" events fire before the "Did*" events,
since each individual page expects the events to come in the correct order.