Fix StackRouter child transition handling (#12)
See MOBLUAPP-2554.

This change fixes a StackRouter bug that was causing transition-time
errors for nested child routers when you try to transition to a nested
child page before transition to the parent page completes. We would
accidentally replace the parent index with the wrong value because we
were comparing to StackActions.CompleteTransition (which is nil). That
code should have read NavigationActions.CompleteTransition, which is the
proper name. This is a lost-in-translation case from React Navigation
where there's also no language/compiler support that would have
protected us against nil indexing.

This issue only manifested while the parent is transitioning because the
affected code path is invoked only when the child is not already part of
the active route.