Pull Rhodium with Rotriever (#13)
Instead of expecting Rhodium to be in CoreGui.RobloxGui.Modules, this
change makes it so that Rhodium is pulled as a dev package with
rotriever.

This facilitate the workflow, since the developer no longer needs to
open a place that already has Rhodium, since it's bundled within the
plugin. Any place can now run the rhodium tests.

While working around there, TestEZ can also be pulled with rotriever so
it's not needed as a submodule anymore.