Rename test files with .roblox.spec.lua suffix (#24)
When adding react-navigation tests, renaming the test file to
*.roblox.spec and creating the new one with the original name
with the react-navigation tests will lose history for the moved file.
The squashed commit of a PR will consider the moved file just created
and the other modified.

By moving the test files first and adding the tests in another PR,
the history is preserved.