Refactor routers to handle new API (#27)
* Refactor routers to handle new API

Previous API is marked as deprecated. When the previous API is used, it
prints out a warning that shows how the API has changed and what can be
done to refactor existing usage to the new version. The warning message
is only printed once per router (with the stack trace of the call).