Add stack navigator overlay rhodium tests (#14)
Storybooks are now also included into the Rhodium tests plugin, so they
can be reused. It's good for both storybooks and rhodium tests, since it
encourages to implement interactable storybooks and it simplifies the
rhodium tests.

To avoid problems caused by XPath when test failures happens, a utility
function was added to return a ScreenGui with a unique name. That way,
if a test fails and do not destroy the ScreenGui, the next test will not
obtain the previous ScreenGui (through XPath) because it has a new name.

The utility `TrackRobloxStackNavigatorRoute` is used to wait for transitions
to finish before asserting that the correct route is shown in Rhodium tests.