Integrate changes from v0.4.1-patch (#97)
# Includes 3 changes:

## Problem
[LUAFDN-386](https://jira.rbx.com/browse/LUAFDN-386) Roact Navigation exception: Listener disconnected twice

The problem can only happen if willUnmount is invoked twice (due to potential reentrancy
issues in legacy Roact).

## Solution
The simple workaround is to `nil` out the disconnect callback so it can never be invoked again.

## Testing
Tested manually in Studio

---

## Problem
[LUAFDN-330](https://jira.rbx.com/browse/LUAFDN-330) Allow transparent background on non-overlay screens

The need to show a 3D avatar scene through many of our screens means that we
are forced to set `overlayEnabled = true` on _all_ screens. Otherwise, Roact Navigation
currently uses an opaque background on the StackNavigator screens.

This also forces us to manage visibility of stack screens manually since Roact Navigaiton
treats them all as transparent overlays (all are visible at once).

## Solution
Allow for non-overlay screens to still use a transparent background.

A new navigation option is being added to ensure that this does not cause any
backwards-compatibility issues. If the `transparent = true` option is specified,
the StackViewCard component will be passed `transparent = true` as well.

If the overlayEnabled option is specified, the screen is always considered transparent
and the new `transparent` option will be ignored.

## Testing
Tested in Studio with additional changes to eliminate the manual screen visibility management
and rely on Roact Navigation to do it all automatically.

---

## Problem
[AVBURST-5355](https://jira.rbx.com/browse/AVBURST-5355) Bottom inset background color doesn't match on iPhone X

The current ClipsDescendants prop values used by RN screens is causing a visual display glitch on screens with global gui insets (like the iPhone X). Normally, AvatarExpeirenceView renders a background frame over the bottom inset area to match the background of the rest of the UI. With RN enabled, this is clipped and the avatar scene shows through instead.

## Solution
Do not clip descendants in the RN screen wrapping components. Pages being migrated from AppRouter did not previously have their descendants clipped so this maintains existing behavior and fixes the iPhone rendering issue above.

## Testing
Tested in Studio with iPhone X emulation