AVBURST-8274 Fix race condition in StackView Transitioner (#111)
## Problem
https://jira.rbx.com/browse/AVBURST-8274

Inside `Transitioner:_startTransition`, a call to `setState` is followed by logic
that builds the next transition props (including which screens to render).
The sequence is not obviously linked, but happens as follows:

1. setState(...)
2. position:setGoal(Otter.instant, ...)
3. on next Heartbeat, Otter instant transition finishes, fires onComplete
4. _onTransitionEnd is called
5. buildTransitionProps is called

If the state update from step 1 has not yet been applied by the time we reach
step 5, stale state will be used to build the transition props which determine
which screens are actually rendered.

## Solution
Switch to the two-argument form of `setState`, where the second argument is
a callback that is invoked only after the state update has actually been applied.

The changes are committed as a series of small refactoring steps with the actual
change in behavior only happening in steps 4 and 5 (review with whitespace
changes disabled to more easily see what was changed).