Use unique key for stack screens rather than index (#114)
When a stack screen is replaced, the new screen's index will change
when the transition completes (it will begin with the index _after_
the old screen, but end up with the same index as the old screen, once
the old screen is removed).

Currenty, the screens to render (in StackViewLayout) use the index
as their key. This changes that to the unique RoactNavigation key for
each screen instance. It avoids an unmount/remount of the new screen
during replacement actions because the key is now stable.

As an example of what currently goes wrong, consider an ItemDetails
transition which replaces one ItemDetails screen with another.
Without this change, the following children are used during transition:
```lua
{
    [1] = CatalogScreen,
    [2] = Old ItemDetails screen,
    [3] = New ItemDetails screen,
}
```

The new screen then does not match keys after the old one is removed (3 vs. 2):
```lua
{
    [1] = CatalogScreen,
    [2] = New ItemDetails screen,
}
```

With this change, the children table is instead:
```lua
{
    [key1] = CatalogScreen,
    [key2] = Old ItemDetails screen,
    [key3] = New ItemDetails screen,
}
```

And after the transition completes:
```lua
{
    [key1] = CatalogScreen,
    [key3] = New ItemDetails screen,
}
```