CLIXBOX-3151 Enable Gamepad for Friends List Page roact navigation part (#135)
When using gamepad nav on console, the `AbsorbInput` component will be
focused(highlighted) as they are `Selectable == true`.

Because of the nested nature of the navigators, there will be 3
`AbsorbInput` components (Navigator/init.lua -> ModalStackNavigator ->
MainStackNavigator) to receive the gamepad/keyboard selections.

We didn't need to consider `Selectable` when we implemented the
`AbsorbInput` mechanism 4 years ago, but now we have to consider turning
it false!


![snapshot](https://github.com/Roblox/roact-navigation/assets/4699687/351bed11-326e-4fb7-8764-62c26fd847f6)

---------