APPFDN-2682 Roact Navigation navigator may reference stale navigation state (#144)
https://roblox.atlassian.net/browse/APPFDN-2682

There is a latent bug in Roact Navigation that causes a crash in very
extreme circumstances. The bug occurs when a navigator re-renders in the
middle of a navigation state change. If a parent navigator receives the
new state change before a descendant navigator receives it, but the
child re-renders (for any reason e.g. screenProps change), then the
child will reference the parent's mutable children table which has since
been overwritten with the new state, thus causing a crash. The "extreme"
circumstances come from having a deep component tree which React must
split up over multiple frames.

The fix is to refactor `getChildNavigation` to no longer reference the
mutable table directly, and instead capture the immutable data member it
wants within a closure.

Test added which repros the bug.