APPFDN-2752 StackViewLayout should respect page transparency property (#146)
https://roblox.atlassian.net/browse/APPFDN-2752

When performing optimizations for a card stack, we check for page
transparency to determine which cards should be rendered. The
transparency check only accounts for overlay-enabled screens, and not
the explicit `transparent` option. This causes issues where screens are
not rendered when they should be. This PR fixes the optimization pass to
include a check for the additional property.