StackViewCard remove default background color/transparency (#148)
When rendering a `StackViewCard`, we currently provide a default card
color for the background. In practice, this default isn't particular
useful beyond initial prototyping since background color is heavily
customized based on theme. Additionally, we don't really make use of
this api in the app since most screens typically render their own
background.

However, the main reason we want to remove this default is to allow
"see-through" cards. Roact Navigation already has support for
transparent cards, but the internals of RN carry additional render logic
so that cards underneath will be rendered along with the topmost
transparent cards. In this case, we want our render logic to treat the
topmost card as "opaque" while still being see-through. This is so we
can see the world space behind our UI.